requirements.txt
langchain==0.0.203
openai==0.27.8
python-dotenv==1.0.0
tiktoken==0.4.0
sentence_transformers==2.2.2
chromadb==0.3.26
pydantic==1.10.9
typer==0.9.0
